feat: updated cli commands to echo, cleaned up query, added populate db functionality

+++ src/ape_cache/_cli.py
@@ -18,7 +18,15 @@ def cli():
 
 @cli.command(short_help="Initialize a new cache db")
 def init():
-    models.Base.metadata.create_all(bind=get_engine().engine)
+    db_file = get_engine().database_file
+    if not db_file.is_file():
+        click.echo("initializing database")
+        models.Base.metadata.create_all(bind=get_engine().engine)
+        return click.echo("caching database initialized")
+    click.echo("caching database already exists!")
+
+
+# TODO: add migrate command
 
 
 @cli.command(short_help="Call and print SQL Statement to the cache db")
@@ -31,4 +39,8 @@ def query(sql):
 @cli.command(short_help="Purges entire database")
 def purge():
     db_file = get_engine().database_file
+    if not db_file.is_file():
+        # Add check here to show we have a file that exists
+        return click.echo("caching database must be initialized with `ape cache init`")
     db_file.unlink()
+    click.echo("caching database purged")
